// src/tasks/enums/task.enum.ts
export enum TaskCategoryEnum {
  disolution = 'Disolution',
  disolution_chatbot = 'Disolution & Chatbot',
  chatbot = 'Chatbot',
}

export enum TaskTypeEnum {
  encore_task = 'encore_task',
  client_task = 'client_task',
}

export enum TaskStatusEnum {
  pending = 'pending',
  in_progress = 'in_progress',
  complete = 'complete',
}